% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca_screen}
\alias{brca_screen}
\title{Project Drive screen data for two brca cell lines}
\format{
A data.table with three columns: "screen_entrez", "screen_symbol" and "cell_line".
}
\usage{
brca_screen
}
\description{
Screen hits are selected with RSA <= -2.
}
\keyword{datasets}
