% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_compSLP}
\alias{example_compSLP}
\title{SLPs predicted by comp_slp}
\format{
A data.table.
}
\usage{
data(example_compSLP)
}
\description{
SLPs predicted by comp_slp
}
\keyword{datasets}
