% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_corr_mut}
\alias{example_corr_mut}
\title{Patients mutations to be use in the corr_slp}
\format{
A data.table.
}
\usage{
data(example_corr_mut)
}
\description{
Mutations and related TCGA ids.
}
\keyword{datasets}
