% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_corrSLP}
\alias{example_corrSLP}
\title{SLPs predicted by corr_slp}
\format{
A data.table.
}
\usage{
data(example_corrSLP)
}
\description{
SLPs predicted by corr_slp
}
\keyword{datasets}
