% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie3.R
\name{getlink}
\alias{getlink}
\title{Get sorted list of regulatory links in GENIE3 results}
\usage{
getlink(weight.matrix, report.max = NULL)
}
\arguments{
\item{weight.matrix}{a weighted adjacency matrix as returned by get.weight.matrix.targ.}

\item{report.max}{maximum number of links to report (default all links).}
}
\value{
A data.table of links with columns "from.gene", "to.gene", "im".
}
\description{
Take genie3 output and sort the links.
}
