% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_expr}
\alias{example_expr}
\title{Expression data to be used in comp_slp}
\format{
A matrix.
}
\usage{
data(example_expr)
}
\description{
Expresion matrix, genes by samples.
}
\keyword{datasets}
